successfully commited task 1

+++ 0x18-dynamic_libraries/0-memset.c
@@ -8,7 +8,7 @@
  * Return: A pointer to the filled memory area @s.
  */
 
-void *_memset(void *s, int c, size_t n)
+void *_memset(void *s, int c , unsigned n)
 {
 	unsigned int index;
 	unsigned char *memory = s, value = c;
